refactor(lsp): move zk markdown config to lsp module
Zk can now be initialize alongside other language servers as there is a
valid entry for it in the mason package list.
